mongo change stream
types

search impl

init typesense

format

fix api & brittler

fix performance

search done

fixes

dev deployment

fix: remove excess args on mongo str

fix: minio creds

fix(data-brittler): some things

fix(data-brittler): entitlements job

fix(data-brittler): emote mod requests

fix(data-brittler): messages job

fix(api): search

fix(api): emote search paging issue

docs(api): add todo

wd

+++ shared/src/database/types/duration.rs
@@ -1,7 +1,7 @@
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
+#[derive(Debug, Clone, Copy, serde::Serialize, serde::Deserialize)]
 #[serde(deny_unknown_fields)]
 #[serde(tag = "unit", content = "count")]
 pub enum DurationUnit {
-	Days(u64),
-	Months(u64),
+	Days(i32),
+	Months(i32),
 }
